170
Update for v1.7.0
       LIBRARY  UPDATES
-------------------------------

Visualise - can be accessed by
Model.visualise()  -> takes time to display , slight time to process
Model.goFigure()
Model.snap()

Model.py
-> Updated the Model.Select.Polygon

View.py
-> Cutting diagram
-> Update to beam diagram (need to check as software is finicky)
- Plate Stress
- Cutting line
- Cutting Plane

Analysis control
- HOH Control

Element
- Solid element sync bug
- GMSH removed for streamlit

Load
- Spec Disp create , clear ,
- Load Case, Group in Nodal Mass

Moving Load
- improved the MV Code application

Offset
- Angle + T shape

Result table
- HOH stress / Temp / disp / tensile_stress / Pipe temp

- TH Disp / Vel / Acc / Beam force / Truss force / Plate force / Solid / Wall / Stress
- Vibration mode shape / Buckling mode shapes

+++ midas_civil/_node.py
@@ -115,9 +115,9 @@ class Node:
         #     node_count = max(Node.ids)+1
         
         
-        self.X = round(x,6)
-        self.Y = round(y,6)
-        self.Z = round(z,6)
+        self.X = float(round(x,6))
+        self.Y = float(round(y,6))
+        self.Z = float(round(z,6))
         self.AXIS = [[0,0,0],[0,0,0],[0,0,0]]
 
         if id == 0 : self.ID = node_count
